third_party/blink/web_tests/platform/generic/editing/execCommand/enabling-and-selection-expected.txt

This tests enabling of execCommand commands based on whether the selection is a caret or range or in editable content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS whenEnabled('FindString') is 'always'
PASS whenEnabled('Print') is 'always'
PASS whenEnabled('SelectAll') is 'always'
PASS whenEnabled('Transpose') is 'caret'
PASS whenEnabled('Copy') is 'range'
PASS whenEnabled('Cut') is 'editable range'
PASS whenEnabled('RemoveFormat') is 'editable range'
PASS whenEnabled('Delete') is 'editable'
PASS whenEnabled('FontName') is 'richly editable'
PASS whenEnabled('FontSize') is 'richly editable'
PASS whenEnabled('FontSizeDelta') is 'richly editable'
PASS whenEnabled('ForwardDelete') is 'editable'
PASS whenEnabled('InsertHTML') is 'editable'
PASS whenEnabled('InsertParagraph') is 'editable'
PASS whenEnabled('InsertText') is 'editable'
PASS whenEnabled('BackColor') is 'richly editable'
PASS whenEnabled('Bold') is 'richly editable'
PASS whenEnabled('CreateLink') is 'richly editable'
PASS whenEnabled('ForeColor') is 'richly editable'
PASS whenEnabled('FormatBlock') is 'richly editable'
PASS whenEnabled('HiliteColor') is 'richly editable'
PASS whenEnabled('Indent') is 'richly editable'
PASS whenEnabled('InsertHorizontalRule') is 'richly editable'
PASS whenEnabled('InsertImage') is 'richly editable'
PASS whenEnabled('InsertNewlineInQuotedContent') is 'richly editable'
PASS whenEnabled('InsertOrderedList') is 'richly editable'
PASS whenEnabled('InsertUnorderedList') is 'richly editable'
PASS whenEnabled('Italic') is 'richly editable'
PASS whenEnabled('JustifyCenter') is 'richly editable'
PASS whenEnabled('JustifyFull') is 'richly editable'
PASS whenEnabled('JustifyLeft') is 'richly editable'
PASS whenEnabled('JustifyNone') is 'richly editable'
PASS whenEnabled('JustifyRight') is 'richly editable'
PASS whenEnabled('Outdent') is 'richly editable'
PASS whenEnabled('Strikethrough') is 'richly editable'
PASS whenEnabled('Subscript') is 'richly editable'
PASS whenEnabled('Superscript') is 'richly editable'
PASS whenEnabled('Underline') is 'richly editable'
PASS whenEnabled('Paste') is 'editable'
PASS whenEnabled('PasteAndMatchStyle') is 'editable'
PASS whenEnabled('Unlink') is 'richly editable range'
PASS whenEnabled('Unselect') is 'visible'
PASS successfullyParsed is true

TEST COMPLETE
